Move classes to a separate header file
