+++ src/avx512-32bit-qsort.hpp
@@ -23,313 +23,6 @@
 #define NETWORK_32BIT_6 11, 10, 9, 8, 15, 14, 13, 12, 3, 2, 1, 0, 7, 6, 5, 4
 #define NETWORK_32BIT_7 7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8
 
-template <>
-struct zmm_vector<int32_t> {
-    using type_t = int32_t;
-    using zmm_t = __m512i;
-    using ymm_t = __m256i;
-    using opmask_t = __mmask16;
-    static const uint8_t numlanes = 16;
-
-    static type_t type_max()
-    {
-        return X86_SIMD_SORT_MAX_INT32;
-    }
-    static type_t type_min()
-    {
-        return X86_SIMD_SORT_MIN_INT32;
-    }
-    static zmm_t zmm_max()
-    {
-        return _mm512_set1_epi32(type_max());
-    }
-
-    static opmask_t knot_opmask(opmask_t x)
-    {
-        return _mm512_knot(x);
-    }
-    static opmask_t ge(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_epi32_mask(x, y, _MM_CMPINT_NLT);
-    }
-    template <int scale>
-    static ymm_t i64gather(__m512i index, void const *base)
-    {
-        return _mm512_i64gather_epi32(index, base, scale);
-    }
-    static zmm_t merge(ymm_t y1, ymm_t y2)
-    {
-        zmm_t z1 = _mm512_castsi256_si512(y1);
-        return _mm512_inserti32x8(z1, y2, 1);
-    }
-    static zmm_t loadu(void const *mem)
-    {
-        return _mm512_loadu_si512(mem);
-    }
-    static void mask_compressstoreu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_compressstoreu_epi32(mem, mask, x);
-    }
-    static zmm_t mask_loadu(zmm_t x, opmask_t mask, void const *mem)
-    {
-        return _mm512_mask_loadu_epi32(x, mask, mem);
-    }
-    static zmm_t mask_mov(zmm_t x, opmask_t mask, zmm_t y)
-    {
-        return _mm512_mask_mov_epi32(x, mask, y);
-    }
-    static void mask_storeu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_storeu_epi32(mem, mask, x);
-    }
-    static zmm_t min(zmm_t x, zmm_t y)
-    {
-        return _mm512_min_epi32(x, y);
-    }
-    static zmm_t max(zmm_t x, zmm_t y)
-    {
-        return _mm512_max_epi32(x, y);
-    }
-    static zmm_t permutexvar(__m512i idx, zmm_t zmm)
-    {
-        return _mm512_permutexvar_epi32(idx, zmm);
-    }
-    static type_t reducemax(zmm_t v)
-    {
-        return _mm512_reduce_max_epi32(v);
-    }
-    static type_t reducemin(zmm_t v)
-    {
-        return _mm512_reduce_min_epi32(v);
-    }
-    static zmm_t set1(type_t v)
-    {
-        return _mm512_set1_epi32(v);
-    }
-    template <uint8_t mask>
-    static zmm_t shuffle(zmm_t zmm)
-    {
-        return _mm512_shuffle_epi32(zmm, (_MM_PERM_ENUM)mask);
-    }
-    static void storeu(void *mem, zmm_t x)
-    {
-        return _mm512_storeu_si512(mem, x);
-    }
-
-    static ymm_t max(ymm_t x, ymm_t y)
-    {
-        return _mm256_max_epi32(x, y);
-    }
-    static ymm_t min(ymm_t x, ymm_t y)
-    {
-        return _mm256_min_epi32(x, y);
-    }
-};
-template <>
-struct zmm_vector<uint32_t> {
-    using type_t = uint32_t;
-    using zmm_t = __m512i;
-    using ymm_t = __m256i;
-    using opmask_t = __mmask16;
-    static const uint8_t numlanes = 16;
-
-    static type_t type_max()
-    {
-        return X86_SIMD_SORT_MAX_UINT32;
-    }
-    static type_t type_min()
-    {
-        return 0;
-    }
-    static zmm_t zmm_max()
-    {
-        return _mm512_set1_epi32(type_max());
-    } // TODO: this should broadcast bits as is?
-
-    template <int scale>
-    static ymm_t i64gather(__m512i index, void const *base)
-    {
-        return _mm512_i64gather_epi32(index, base, scale);
-    }
-    static zmm_t merge(ymm_t y1, ymm_t y2)
-    {
-        zmm_t z1 = _mm512_castsi256_si512(y1);
-        return _mm512_inserti32x8(z1, y2, 1);
-    }
-    static opmask_t knot_opmask(opmask_t x)
-    {
-        return _mm512_knot(x);
-    }
-    static opmask_t ge(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_epu32_mask(x, y, _MM_CMPINT_NLT);
-    }
-    static zmm_t loadu(void const *mem)
-    {
-        return _mm512_loadu_si512(mem);
-    }
-    static zmm_t max(zmm_t x, zmm_t y)
-    {
-        return _mm512_max_epu32(x, y);
-    }
-    static void mask_compressstoreu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_compressstoreu_epi32(mem, mask, x);
-    }
-    static zmm_t mask_loadu(zmm_t x, opmask_t mask, void const *mem)
-    {
-        return _mm512_mask_loadu_epi32(x, mask, mem);
-    }
-    static zmm_t mask_mov(zmm_t x, opmask_t mask, zmm_t y)
-    {
-        return _mm512_mask_mov_epi32(x, mask, y);
-    }
-    static void mask_storeu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_storeu_epi32(mem, mask, x);
-    }
-    static zmm_t min(zmm_t x, zmm_t y)
-    {
-        return _mm512_min_epu32(x, y);
-    }
-    static zmm_t permutexvar(__m512i idx, zmm_t zmm)
-    {
-        return _mm512_permutexvar_epi32(idx, zmm);
-    }
-    static type_t reducemax(zmm_t v)
-    {
-        return _mm512_reduce_max_epu32(v);
-    }
-    static type_t reducemin(zmm_t v)
-    {
-        return _mm512_reduce_min_epu32(v);
-    }
-    static zmm_t set1(type_t v)
-    {
-        return _mm512_set1_epi32(v);
-    }
-    template <uint8_t mask>
-    static zmm_t shuffle(zmm_t zmm)
-    {
-        return _mm512_shuffle_epi32(zmm, (_MM_PERM_ENUM)mask);
-    }
-    static void storeu(void *mem, zmm_t x)
-    {
-        return _mm512_storeu_si512(mem, x);
-    }
-
-    static ymm_t max(ymm_t x, ymm_t y)
-    {
-        return _mm256_max_epu32(x, y);
-    }
-    static ymm_t min(ymm_t x, ymm_t y)
-    {
-        return _mm256_min_epu32(x, y);
-    }
-};
-template <>
-struct zmm_vector<float> {
-    using type_t = float;
-    using zmm_t = __m512;
-    using ymm_t = __m256;
-    using opmask_t = __mmask16;
-    static const uint8_t numlanes = 16;
-
-    static type_t type_max()
-    {
-        return X86_SIMD_SORT_INFINITYF;
-    }
-    static type_t type_min()
-    {
-        return -X86_SIMD_SORT_INFINITYF;
-    }
-    static zmm_t zmm_max()
-    {
-        return _mm512_set1_ps(type_max());
-    }
-
-    static opmask_t knot_opmask(opmask_t x)
-    {
-        return _mm512_knot(x);
-    }
-    static opmask_t ge(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_ps_mask(x, y, _CMP_GE_OQ);
-    }
-    template <int scale>
-    static ymm_t i64gather(__m512i index, void const *base)
-    {
-        return _mm512_i64gather_ps(index, base, scale);
-    }
-    static zmm_t merge(ymm_t y1, ymm_t y2)
-    {
-        zmm_t z1 = _mm512_castsi512_ps(
-                _mm512_castsi256_si512(_mm256_castps_si256(y1)));
-        return _mm512_insertf32x8(z1, y2, 1);
-    }
-    static zmm_t loadu(void const *mem)
-    {
-        return _mm512_loadu_ps(mem);
-    }
-    static zmm_t max(zmm_t x, zmm_t y)
-    {
-        return _mm512_max_ps(x, y);
-    }
-    static void mask_compressstoreu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_compressstoreu_ps(mem, mask, x);
-    }
-    static zmm_t mask_loadu(zmm_t x, opmask_t mask, void const *mem)
-    {
-        return _mm512_mask_loadu_ps(x, mask, mem);
-    }
-    static zmm_t mask_mov(zmm_t x, opmask_t mask, zmm_t y)
-    {
-        return _mm512_mask_mov_ps(x, mask, y);
-    }
-    static void mask_storeu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_storeu_ps(mem, mask, x);
-    }
-    static zmm_t min(zmm_t x, zmm_t y)
-    {
-        return _mm512_min_ps(x, y);
-    }
-    static zmm_t permutexvar(__m512i idx, zmm_t zmm)
-    {
-        return _mm512_permutexvar_ps(idx, zmm);
-    }
-    static type_t reducemax(zmm_t v)
-    {
-        return _mm512_reduce_max_ps(v);
-    }
-    static type_t reducemin(zmm_t v)
-    {
-        return _mm512_reduce_min_ps(v);
-    }
-    static zmm_t set1(type_t v)
-    {
-        return _mm512_set1_ps(v);
-    }
-    template <uint8_t mask>
-    static zmm_t shuffle(zmm_t zmm)
-    {
-        return _mm512_shuffle_ps(zmm, zmm, (_MM_PERM_ENUM)mask);
-    }
-    static void storeu(void *mem, zmm_t x)
-    {
-        return _mm512_storeu_ps(mem, x);
-    }
-
-    static ymm_t max(ymm_t x, ymm_t y)
-    {
-        return _mm256_max_ps(x, y);
-    }
-    static ymm_t min(ymm_t x, ymm_t y)
-    {
-        return _mm256_min_ps(x, y);
-    }
-};
 
 /*
  * Assumes zmm is random and performs a full sorting network defined in
