+++ src/avx512-64bit-common.h
@@ -19,322 +19,6 @@
 #define NETWORK_64BIT_3 5, 4, 7, 6, 1, 0, 3, 2
 #define NETWORK_64BIT_4 3, 2, 1, 0, 7, 6, 5, 4
 
-template <>
-struct zmm_vector<int64_t> {
-    using type_t = int64_t;
-    using zmm_t = __m512i;
-    using ymm_t = __m512i;
-    using opmask_t = __mmask8;
-    static const uint8_t numlanes = 8;
-
-    static type_t type_max()
-    {
-        return X86_SIMD_SORT_MAX_INT64;
-    }
-    static type_t type_min()
-    {
-        return X86_SIMD_SORT_MIN_INT64;
-    }
-    static zmm_t zmm_max()
-    {
-        return _mm512_set1_epi64(type_max());
-    } // TODO: this should broadcast bits as is?
-
-    static zmm_t set(type_t v1,
-                     type_t v2,
-                     type_t v3,
-                     type_t v4,
-                     type_t v5,
-                     type_t v6,
-                     type_t v7,
-                     type_t v8)
-    {
-        return _mm512_set_epi64(v1, v2, v3, v4, v5, v6, v7, v8);
-    }
-
-    static opmask_t knot_opmask(opmask_t x)
-    {
-        return _knot_mask8(x);
-    }
-    static opmask_t ge(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_epi64_mask(x, y, _MM_CMPINT_NLT);
-    }
-    static opmask_t eq(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_epi64_mask(x, y, _MM_CMPINT_EQ);
-    }
-    template <int scale>
-    static zmm_t i64gather(__m512i index, void const *base)
-    {
-        return _mm512_i64gather_epi64(index, base, scale);
-    }
-    static zmm_t loadu(void const *mem)
-    {
-        return _mm512_loadu_si512(mem);
-    }
-    static zmm_t max(zmm_t x, zmm_t y)
-    {
-        return _mm512_max_epi64(x, y);
-    }
-    static void mask_compressstoreu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_compressstoreu_epi64(mem, mask, x);
-    }
-    static zmm_t mask_loadu(zmm_t x, opmask_t mask, void const *mem)
-    {
-        return _mm512_mask_loadu_epi64(x, mask, mem);
-    }
-    static zmm_t mask_mov(zmm_t x, opmask_t mask, zmm_t y)
-    {
-        return _mm512_mask_mov_epi64(x, mask, y);
-    }
-    static void mask_storeu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_storeu_epi64(mem, mask, x);
-    }
-    static zmm_t min(zmm_t x, zmm_t y)
-    {
-        return _mm512_min_epi64(x, y);
-    }
-    static zmm_t permutexvar(__m512i idx, zmm_t zmm)
-    {
-        return _mm512_permutexvar_epi64(idx, zmm);
-    }
-    static type_t reducemax(zmm_t v)
-    {
-        return _mm512_reduce_max_epi64(v);
-    }
-    static type_t reducemin(zmm_t v)
-    {
-        return _mm512_reduce_min_epi64(v);
-    }
-    static zmm_t set1(type_t v)
-    {
-        return _mm512_set1_epi64(v);
-    }
-    template <uint8_t mask>
-    static zmm_t shuffle(zmm_t zmm)
-    {
-        __m512d temp = _mm512_castsi512_pd(zmm);
-        return _mm512_castpd_si512(
-                _mm512_shuffle_pd(temp, temp, (_MM_PERM_ENUM)mask));
-    }
-    static void storeu(void *mem, zmm_t x)
-    {
-        return _mm512_storeu_si512(mem, x);
-    }
-};
-template <>
-struct zmm_vector<uint64_t> {
-    using type_t = uint64_t;
-    using zmm_t = __m512i;
-    using ymm_t = __m512i;
-    using opmask_t = __mmask8;
-    static const uint8_t numlanes = 8;
-
-    static type_t type_max()
-    {
-        return X86_SIMD_SORT_MAX_UINT64;
-    }
-    static type_t type_min()
-    {
-        return 0;
-    }
-    static zmm_t zmm_max()
-    {
-        return _mm512_set1_epi64(type_max());
-    }
-
-    static zmm_t set(type_t v1,
-                     type_t v2,
-                     type_t v3,
-                     type_t v4,
-                     type_t v5,
-                     type_t v6,
-                     type_t v7,
-                     type_t v8)
-    {
-        return _mm512_set_epi64(v1, v2, v3, v4, v5, v6, v7, v8);
-    }
-
-    template <int scale>
-    static zmm_t i64gather(__m512i index, void const *base)
-    {
-        return _mm512_i64gather_epi64(index, base, scale);
-    }
-    static opmask_t knot_opmask(opmask_t x)
-    {
-        return _knot_mask8(x);
-    }
-    static opmask_t ge(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_epu64_mask(x, y, _MM_CMPINT_NLT);
-    }
-    static opmask_t eq(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_epu64_mask(x, y, _MM_CMPINT_EQ);
-    }
-    static zmm_t loadu(void const *mem)
-    {
-        return _mm512_loadu_si512(mem);
-    }
-    static zmm_t max(zmm_t x, zmm_t y)
-    {
-        return _mm512_max_epu64(x, y);
-    }
-    static void mask_compressstoreu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_compressstoreu_epi64(mem, mask, x);
-    }
-    static zmm_t mask_loadu(zmm_t x, opmask_t mask, void const *mem)
-    {
-        return _mm512_mask_loadu_epi64(x, mask, mem);
-    }
-    static zmm_t mask_mov(zmm_t x, opmask_t mask, zmm_t y)
-    {
-        return _mm512_mask_mov_epi64(x, mask, y);
-    }
-    static void mask_storeu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_storeu_epi64(mem, mask, x);
-    }
-    static zmm_t min(zmm_t x, zmm_t y)
-    {
-        return _mm512_min_epu64(x, y);
-    }
-    static zmm_t permutexvar(__m512i idx, zmm_t zmm)
-    {
-        return _mm512_permutexvar_epi64(idx, zmm);
-    }
-    static type_t reducemax(zmm_t v)
-    {
-        return _mm512_reduce_max_epu64(v);
-    }
-    static type_t reducemin(zmm_t v)
-    {
-        return _mm512_reduce_min_epu64(v);
-    }
-    static zmm_t set1(type_t v)
-    {
-        return _mm512_set1_epi64(v);
-    }
-    template <uint8_t mask>
-    static zmm_t shuffle(zmm_t zmm)
-    {
-        __m512d temp = _mm512_castsi512_pd(zmm);
-        return _mm512_castpd_si512(
-                _mm512_shuffle_pd(temp, temp, (_MM_PERM_ENUM)mask));
-    }
-    static void storeu(void *mem, zmm_t x)
-    {
-        return _mm512_storeu_si512(mem, x);
-    }
-};
-template <>
-struct zmm_vector<double> {
-    using type_t = double;
-    using zmm_t = __m512d;
-    using ymm_t = __m512d;
-    using opmask_t = __mmask8;
-    static const uint8_t numlanes = 8;
-
-    static type_t type_max()
-    {
-        return X86_SIMD_SORT_INFINITY;
-    }
-    static type_t type_min()
-    {
-        return -X86_SIMD_SORT_INFINITY;
-    }
-    static zmm_t zmm_max()
-    {
-        return _mm512_set1_pd(type_max());
-    }
-
-    static zmm_t set(type_t v1,
-                     type_t v2,
-                     type_t v3,
-                     type_t v4,
-                     type_t v5,
-                     type_t v6,
-                     type_t v7,
-                     type_t v8)
-    {
-        return _mm512_set_pd(v1, v2, v3, v4, v5, v6, v7, v8);
-    }
-
-    static opmask_t knot_opmask(opmask_t x)
-    {
-        return _knot_mask8(x);
-    }
-    static opmask_t ge(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_pd_mask(x, y, _CMP_GE_OQ);
-    }
-    static opmask_t eq(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_pd_mask(x, y, _CMP_EQ_OQ);
-    }
-    template <int scale>
-    static zmm_t i64gather(__m512i index, void const *base)
-    {
-        return _mm512_i64gather_pd(index, base, scale);
-    }
-    static zmm_t loadu(void const *mem)
-    {
-        return _mm512_loadu_pd(mem);
-    }
-    static zmm_t max(zmm_t x, zmm_t y)
-    {
-        return _mm512_max_pd(x, y);
-    }
-    static void mask_compressstoreu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_compressstoreu_pd(mem, mask, x);
-    }
-    static zmm_t mask_loadu(zmm_t x, opmask_t mask, void const *mem)
-    {
-        return _mm512_mask_loadu_pd(x, mask, mem);
-    }
-    static zmm_t mask_mov(zmm_t x, opmask_t mask, zmm_t y)
-    {
-        return _mm512_mask_mov_pd(x, mask, y);
-    }
-    static void mask_storeu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_storeu_pd(mem, mask, x);
-    }
-    static zmm_t min(zmm_t x, zmm_t y)
-    {
-        return _mm512_min_pd(x, y);
-    }
-    static zmm_t permutexvar(__m512i idx, zmm_t zmm)
-    {
-        return _mm512_permutexvar_pd(idx, zmm);
-    }
-    static type_t reducemax(zmm_t v)
-    {
-        return _mm512_reduce_max_pd(v);
-    }
-    static type_t reducemin(zmm_t v)
-    {
-        return _mm512_reduce_min_pd(v);
-    }
-    static zmm_t set1(type_t v)
-    {
-        return _mm512_set1_pd(v);
-    }
-    template <uint8_t mask>
-    static zmm_t shuffle(zmm_t zmm)
-    {
-        return _mm512_shuffle_pd(zmm, zmm, (_MM_PERM_ENUM)mask);
-    }
-    static void storeu(void *mem, zmm_t x)
-    {
-        return _mm512_storeu_pd(mem, x);
-    }
-};
 X86_SIMD_SORT_INLINE int64_t replace_nan_with_inf(double *arr, int64_t arrsize)
 {
     int64_t nan_count = 0;
