+++ src/avx512fp16-16bit-qsort.hpp
@@ -9,111 +9,6 @@
 
 #include "avx512-16bit-common.h"
 
-typedef union {
-    _Float16 f_;
-    uint16_t i_;
-} Fp16Bits;
-
-template <>
-struct zmm_vector<_Float16> {
-    using type_t = _Float16;
-    using zmm_t = __m512h;
-    using ymm_t = __m256h;
-    using opmask_t = __mmask32;
-    static const uint8_t numlanes = 32;
-
-    static __m512i get_network(int index)
-    {
-        return _mm512_loadu_si512(&network[index - 1][0]);
-    }
-    static type_t type_max()
-    {
-        Fp16Bits val;
-        val.i_ = X86_SIMD_SORT_INFINITYH;
-        return val.f_;
-    }
-    static type_t type_min()
-    {
-        Fp16Bits val;
-        val.i_ = X86_SIMD_SORT_NEGINFINITYH;
-        return val.f_;
-    }
-    static zmm_t zmm_max()
-    {
-        return _mm512_set1_ph(type_max());
-    }
-    static opmask_t knot_opmask(opmask_t x)
-    {
-        return _knot_mask32(x);
-    }
-
-    static opmask_t ge(zmm_t x, zmm_t y)
-    {
-        return _mm512_cmp_ph_mask(x, y, _CMP_GE_OQ);
-    }
-    static zmm_t loadu(void const *mem)
-    {
-        return _mm512_loadu_ph(mem);
-    }
-    static zmm_t max(zmm_t x, zmm_t y)
-    {
-        return _mm512_max_ph(x, y);
-    }
-    static void mask_compressstoreu(void *mem, opmask_t mask, zmm_t x)
-    {
-        __m512i temp = _mm512_castph_si512(x);
-        // AVX512_VBMI2
-        return _mm512_mask_compressstoreu_epi16(mem, mask, temp);
-    }
-    static zmm_t mask_loadu(zmm_t x, opmask_t mask, void const *mem)
-    {
-        // AVX512BW
-        return _mm512_castsi512_ph(
-                _mm512_mask_loadu_epi16(_mm512_castph_si512(x), mask, mem));
-    }
-    static zmm_t mask_mov(zmm_t x, opmask_t mask, zmm_t y)
-    {
-        return _mm512_castsi512_ph(_mm512_mask_mov_epi16(
-                _mm512_castph_si512(x), mask, _mm512_castph_si512(y)));
-    }
-    static void mask_storeu(void *mem, opmask_t mask, zmm_t x)
-    {
-        return _mm512_mask_storeu_epi16(mem, mask, _mm512_castph_si512(x));
-    }
-    static zmm_t min(zmm_t x, zmm_t y)
-    {
-        return _mm512_min_ph(x, y);
-    }
-    static zmm_t permutexvar(__m512i idx, zmm_t zmm)
-    {
-        return _mm512_permutexvar_ph(idx, zmm);
-    }
-    static type_t reducemax(zmm_t v)
-    {
-        return _mm512_reduce_max_ph(v);
-    }
-    static type_t reducemin(zmm_t v)
-    {
-        return _mm512_reduce_min_ph(v);
-    }
-    static zmm_t set1(type_t v)
-    {
-        return _mm512_set1_ph(v);
-    }
-    template <uint8_t mask>
-    static zmm_t shuffle(zmm_t zmm)
-    {
-        __m512i temp = _mm512_shufflehi_epi16(_mm512_castph_si512(zmm),
-                                              (_MM_PERM_ENUM)mask);
-        return _mm512_castsi512_ph(
-                _mm512_shufflelo_epi16(temp, (_MM_PERM_ENUM)mask));
-    }
-    static void storeu(void *mem, zmm_t x)
-    {
-        return _mm512_storeu_ph(mem, x);
-    }
-};
-
 X86_SIMD_SORT_INLINE int64_t replace_nan_with_inf(_Float16 *arr,
                                                   int64_t arrsize)
 {
